requirement.txt
opencv-python
ultralytics
easyocr
pyserial
reportlab
numpy
lapx
